% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSmooth.R
\name{construct_network}
\alias{construct_network}
\title{Builds a graph using a list of interacting genes or proteins}
\usage{
construct_network(edges, type = "laplacian")
}
\arguments{
\item{edges}{a matrix or a data.frame with two columns listing pairs of interacting proteins or genes (column1 corresponds to protein1 and column2 corresponds to protein2)}

\item{type}{Type of graph that is created from the protein interaction list. By default: laplacian matrix is generated where each edge is normalized by the degree of interacting proteins. If the proteinA interacts with proteinB it gives 1 & upon normalization 1 will be divided by the square root of product of degrees of proteinA and proteinB.}
}
\value{
Graph created from the edge list (interacting proteins)
}
\description{
construct_network is called within network_mapping for generating a laplacian matrix of graph from the protein interaction list
}
\examples{

}

