% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSmooth.R
\name{construct_weighted_network}
\alias{construct_weighted_network}
\title{Builds an edge weighted graph using a list of interacting genes or proteins}
\usage{
construct_weighted_network(edges, type = "laplacian", weights)
}
\arguments{
\item{edges}{a matrix or a data.frame with two columns listing pairs of interacting proteins or genes (column1 corresponds to protein1 and column2 corresponds to protein2)}

\item{type}{Type of graph that is created from the protein interaction list. By default: laplacian matrix is generated where each edge is normalized by the degree of interacting proteins. If the proteinA interacts with proteinB it gives 1 & upon normalization 1 will be divided by the square root of product of degrees of proteinA and proteinB.}

\item{weights}{numeric vector representing the weights of the edges. If the source of protein interactions is STRING, then combined score can be used as weights by dividing them with 1000. Typically weights gives the confidence of the interaction between nodes and ranges from 0 to 1, 1 being the 100 percent confidence of that the two proteins are interacting in the species.}
}
\value{
Edge weighted graph created from the protein interaction list.
}
\description{
construct_weighted_network is called within network_mapping_weighted for generating a edge weighted laplacian matrix of graph from the protein interaction list
}
\examples{

}

