% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetSmooth.R
\name{network_mapping}
\alias{network_mapping}
\title{Maps initial expression scores to an interaction network}
\usage{
network_mapping(network, expr_mat, type = "laplacian", merge.by = "gene_id",
  global = TRUE)
}
\arguments{
\item{network}{species specific protein protein interaction list (typically first two columns:protein1 & protein2) from STRING database or species specific gene interaction list from any desired database.}

\item{expr_mat}{gene expression measures (absolute log2 fold changes or negative log p-values) or mutation profiles or any input values in a dataframe or matrix format. First column of expr_mat should contain gene id and other columns corresponding to other conditions (samples) that has to be propagated on the network.}

\item{type}{Type of graph that is created from the protein interaction list. By default: laplacian matrix is generated where each edge is normalized by the degree of interacting proteins.}

\item{merge.by}{string specifying a character vector in expr_mat (usually the first column name) that is matching with gene or protein names in interaction list}

\item{global}{TRUE indicates that all the input values of a condition should be mapped if the corresponding nodes are present in network. The nodes which doesnot have input values are assigned with 0 as their starting value.}
}
\value{
list of laplacian matrix from the protein interaction list, intensity (initial scores) matrix matching to the nodes in network and node ids as gene_names
}
\description{
This combines gene expression of various conditions with interaction network
}
\examples{
data (expr_mat)
data (network)

# To map the p-values from differential expression analysis to the STRING interaction network
netmap <- network_mapping(network = network, expr_mat = expr_mat, type = "laplacian", merge.by = "gene_id", global = TRUE)

}

